Player labels
Added a TextMeshProUGUI prefab and made each player have one.
In order for it to be reasonable, it is parented to the canvas since it is a UI object.

The rendering is not perfect, it is kind of fuzzy unless I render during fixedupdate, and even so the remote players label seems fuzzy.