Second tilemap and sort order
Adds a second tilemap L1 which is at sort order 2, which is the same as the players
The first tilemap is now sort order 1.

The four  server sidec layers are then mapped to these two tilemaps.

The two upper layers are pretty much the objects that the player can go in front
and behind. The effect is not perfect but is progress,