rescale game
Player is now 1.0 scale,  down from 2.0 from two days ago. Therefore the default camera scale is now 0.5 which we now set at the start of playerController.

This causes two things: the player is now about the size of a tile so there are no more clipping issues with the tiles of layers 2 or 3. Yay. Second, the world is larger which makes it more realistic, wrt to structures.

Now the player moves a bit too fast but that will be fixed server-side.