Version 20.12.27a
One logical fix on the way the z order was assigned, in particular taking into account the temporary z level boost the server does while traversing stairs.

Otherwise we renumber the sorting order to reflect the server sense and the maps
as well. The ground is -1. the players start at 0 (z_level) and go from there.