refactor local vs remote player
A big refactor with some extra changes that could have gone independently.

The meat is the separation of PlayerManager into a base class and two subclasses
for remote and local player. In addition better encapsulation of UI (hud) changes.

The other change is the addtion of a health member to player and the basic network support for it, which is mostly untested.

This updates the protocol to 115/61.